requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
drf-spectacular>=0.27
psycopg2-binary>=2.9
python-dotenv>=1.0
django-filter>=23.5
django-cors-headers>=4.3
redis>=5.0
